regression-test :: move answer-file generation into a separate class
